team / social icons / sign in files moved
