+++ frontend/src/components/SignInRegister/SideBar.css
@@ -6,6 +6,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
+
 }
 
 .SideBar-menu-bars {
@@ -14,7 +15,7 @@
 
 /*SIDE NAV LOGGED IN*/
 .SideBar-nav-menu-user-logged {
-    background-image: url("../images/bg-graphics/graphics/dots.png");
+    background-image: url("./dots.png");
     background-color: rgba(0,0,0, 1);
     width: 20vw;
     height: 100vh;
@@ -23,7 +24,7 @@
     position: absolute;
     padding-left: 2vw;
     padding-top: 1vw;
-    z-index: 2000;
+    z-index: 10000;
     top: 6.5vh;
     left: -100vw;
     transition: 850ms;
@@ -37,7 +38,7 @@
 /* SIDE NAV NOT LOGGED IN*/
 
 .SideBar-nav-menu-user-not-logged {
-    background-image: url("../images/bg-graphics/graphics/dots.png");
+    background-image: url("./dots.png");
     /*background-color: rgba(255,255,255, .7);*/
     /*background-color: rgba(67,13,10, .8);*/
     background-color: rgba(0,0,0, .7);
@@ -48,7 +49,7 @@
     position: absolute;
     padding: 2%;
     /*position: fixed;*/
-    z-index: 2000;
+    z-index: 10000;
     top: 12.5vh;
     right: -500%;
     transition: 850ms;
@@ -92,6 +93,7 @@
 .SideBar-nav-menu-items {
     /*margin-top: 1vh;*/
     width: 100%;
+
 }
 
 .SideBar_Name{
