+++ frontend/src/components/SignInRegister/dots.png
