+++ frontend/src/components/SocialIcons/animations.css
