+++ frontend/src/components/Team/TeamHero.css
@@ -1,0 +1,5 @@
+
+.TeamScreen_hero-img{
+    background-image: url("./images/team.png");
+    background-position: 50% 55%;
+}
