+++ frontend/src/components/Team/images/cody_strack.png
