+++ frontend/src/components/Team/images/deena_smith.png
