+++ frontend/src/components/Team/images/jd_phillips.png
