+++ frontend/src/components/Team/images/jeffrey_rossell.png
