+++ frontend/src/components/Team/images/jerry_fitzgerald.png
