+++ frontend/src/components/Team/images/john_atkin.png
