+++ frontend/src/components/Team/images/john_senecal.png
