+++ frontend/src/components/Team/images/large-team.jpg
