+++ frontend/src/components/Team/images/leslie_banner.png
