+++ frontend/src/components/Team/images/mackeral.png
