+++ frontend/src/components/Team/images/shelby_polasek.png
