+++ frontend/src/components/Team/images/small-team.jpg
