+++ frontend/src/components/Team/images/team.png
