+++ frontend/src/components/Terms/TermsFlir.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import {Container, Row, Col, Form, Button, Nav} from "react-bootstrap";
+import {Container, Row, Col} from "react-bootstrap";
 import {LinkContainer} from "react-router-bootstrap";
