--- frontend/src/style/2/teamScreen/TeamHero.css
@@ -1,5 +1,0 @@
-
-.TeamScreen_hero-img{
-    background-image: url("../../../images/bg-graphics/screen-team/team.png");
-    background-position: 50% 55%;
-}
